refactor: moves options to `types` file
